requirements.txt
Flask==2.3.3
Flask-Session==0.5.0
Flask-CORS==4.0.0
langchain==0.1.20
langchain-openai==0.1.1
langchain-community==0.0.38
pinecone==5.0.0
PyPDF2==3.0.1
python-dotenv==1.0.0
openai==1.35.7
pydantic==2.5.0
Werkzeug==2.3.7
requests==2.31.0
blinker>=1.6.2
itsdangerous>=2.1.2
cachelib>=0.1
